+++ lib/src/theme/themes/default_theme_variant.dart
@@ -65,7 +65,6 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
       decoration:
           ShadDecoration(border: ShadBorder.all(radius: radius, width: 0)),
       gap: 8,
-      iconSize: const Size.square(16),
       expands: false,
     );
   }
@@ -80,7 +79,6 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
       decoration:
           ShadDecoration(border: ShadBorder.all(radius: radius, width: 0)),
       gap: 8,
-      iconSize: const Size.square(16),
       expands: false,
     );
   }
@@ -95,7 +93,6 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
       decoration:
           ShadDecoration(border: ShadBorder.all(radius: radius, width: 0)),
       gap: 8,
-      iconSize: const Size.square(16),
       expands: false,
     );
   }
@@ -114,7 +111,6 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
         ),
       ),
       gap: 8,
-      iconSize: const Size.square(16),
       expands: false,
     );
   }
@@ -128,7 +124,6 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
       decoration:
           ShadDecoration(border: ShadBorder.all(radius: radius, width: 0)),
       gap: 8,
-      iconSize: const Size.square(16),
       expands: false,
     );
   }
@@ -150,19 +145,23 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
       regular: ShadButtonSizeTheme(
         height: 40,
         padding: EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+        iconSize: Size.square(16),
       ),
       sm: ShadButtonSizeTheme(
         height: 36,
         padding: EdgeInsets.symmetric(horizontal: 12),
+        iconSize: Size.square(16),
       ),
       lg: ShadButtonSizeTheme(
         height: 44,
         padding: EdgeInsets.symmetric(horizontal: 32, vertical: 8),
+        iconSize: Size.square(16),
       ),
       icon: ShadButtonSizeTheme(
         height: 40,
         width: 40,
         padding: EdgeInsets.zero,
+        iconSize: Size.square(16),
       ),
     );
   }