+++ playground/lib/pages/button.dart
@@ -59,7 +59,10 @@ class ButtonPage extends StatelessWidget {
             ),
           PlagroundButtonStyle.loading => ShadButton(
               onPressed: () {},
-              icon: const CircularProgressIndicator(strokeWidth: 2),
+              icon: const SizedBox.square(
+                dimension: 16,
+                child: CircularProgressIndicator(strokeWidth: 2),
+              ),
               child: const Text('Please wait'),
             ),
           PlagroundButtonStyle.gradientShadow => ShadButton(